final app: show the info about payoff

+++ real/pages.py
@@ -36,8 +36,8 @@ class questions_graph_01(Page):
                 }
     def before_next_page(self):
         self.player.n_scene = random.randint(1, 8) # escoger el escenario
-        self.player.amount_cultivo_1 = (Constants.cultivo_1[self.player.n_scene-1] * self.player.has_cultivo_1)/10 # calcula el $$$ por el cultivo A
-        self.player.amount_cultivo_2 = (Constants.cultivo_2[self.player.n_scene-1] * self.player.has_cultivo_2)/10 # calcula el $$$ por el cultivo B
+        self.player.amount_cultivo_1 = int((Constants.cultivo_1[self.player.n_scene-1] * self.player.has_cultivo_1)/10) # calcula el $$$ por el cultivo A
+        self.player.amount_cultivo_2 = int((Constants.cultivo_2[self.player.n_scene-1] * self.player.has_cultivo_2)/10) # calcula el $$$ por el cultivo B
         self.player.payoff_amount_app = self.player.amount_cultivo_1 + self.player.amount_cultivo_2
